// cargodelivery-front/generated/models/OrderIn.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cargo delivery
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface OrderIn
 */
export interface OrderIn {
    /**
     * 
     * @type {string}
     * @memberof OrderIn
     */
    userId: string;
    /**
     * 
     * @type {string}
     * @memberof OrderIn
     */
    departureSettlementId: string;
    /**
     * 
     * @type {string}
     * @memberof OrderIn
     */
    destinationSettlementId: string;
    /**
     * 
     * @type {number}
     * @memberof OrderIn
     */
    weight: number;
    /**
     * 
     * @type {number}
     * @memberof OrderIn
     */
    volume: number;
    /**
     * 
     * @type {number}
     * @memberof OrderIn
     */
    price: number;
}

/**
 * Check if a given object implements the OrderIn interface.
 */
export function instanceOfOrderIn(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "userId" in value;
    isInstance = isInstance && "departureSettlementId" in value;
    isInstance = isInstance && "destinationSettlementId" in value;
    isInstance = isInstance && "weight" in value;
    isInstance = isInstance && "volume" in value;
    isInstance = isInstance && "price" in value;

    return isInstance;
}

export function OrderInFromJSON(json: any): OrderIn {
    return OrderInFromJSONTyped(json, false);
}

export function OrderInFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderIn {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'userId': json['userId'],
        'departureSettlementId': json['departureSettlementId'],
        'destinationSettlementId': json['destinationSettlementId'],
        'weight': json['weight'],
        'volume': json['volume'],
        'price': json['price'],
    };
}

export function OrderInToJSON(value?: OrderIn | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'userId': value.userId,
        'departureSettlementId': value.departureSettlementId,
        'destinationSettlementId': value.destinationSettlementId,
        'weight': value.weight,
        'volume': value.volume,
        'price': value.price,
    };
}